+++ twilio/rest/chat/v2/service/__init__.py
@@ -308,7 +308,8 @@ class ServiceContext(InstanceContext):
                pre_webhook_url=values.unset, post_webhook_url=values.unset,
                webhook_method=values.unset, webhook_filters=values.unset,
                limits_channel_members=values.unset,
-               limits_user_channels=values.unset):
+               limits_user_channels=values.unset,
+               media_compatibility_message=values.unset):
         """
         Update the ServiceInstance
 
@@ -339,6 +340,7 @@ class ServiceContext(InstanceContext):
         :param unicode webhook_filters: The webhook_filters
         :param unicode limits_channel_members: The limits.channel_members
         :param unicode limits_user_channels: The limits.user_channels
+        :param unicode media_compatibility_message: The media.compatibility_message
 
         :returns: Updated ServiceInstance
         :rtype: twilio.rest.chat.v2.service.ServiceInstance
@@ -371,6 +373,7 @@ class ServiceContext(InstanceContext):
             'WebhookFilters': webhook_filters,
             'Limits.ChannelMembers': limits_channel_members,
             'Limits.UserChannels': limits_user_channels,
+            'Media.CompatibilityMessage': media_compatibility_message,
         })
 
         payload = self._version.update(
@@ -473,6 +476,7 @@ class ServiceInstance(InstanceResource):
             'webhook_method': payload['webhook_method'],
             'webhook_filters': payload['webhook_filters'],
             'notifications': payload['notifications'],
+            'media': payload['media'],
             'url': payload['url'],
             'links': payload['links'],
         }
@@ -644,6 +648,14 @@ class ServiceInstance(InstanceResource):
         return self._properties['notifications']
 
     @property
+    def media(self):
+        """
+        :returns: The media
+        :rtype: dict
+        """
+        return self._properties['media']
+
+    @property
     def url(self):
         """
         :returns: The url
@@ -700,7 +712,8 @@ class ServiceInstance(InstanceResource):
                pre_webhook_url=values.unset, post_webhook_url=values.unset,
                webhook_method=values.unset, webhook_filters=values.unset,
                limits_channel_members=values.unset,
-               limits_user_channels=values.unset):
+               limits_user_channels=values.unset,
+               media_compatibility_message=values.unset):
         """
         Update the ServiceInstance
 
@@ -731,6 +744,7 @@ class ServiceInstance(InstanceResource):
         :param unicode webhook_filters: The webhook_filters
         :param unicode limits_channel_members: The limits.channel_members
         :param unicode limits_user_channels: The limits.user_channels
+        :param unicode media_compatibility_message: The media.compatibility_message
 
         :returns: Updated ServiceInstance
         :rtype: twilio.rest.chat.v2.service.ServiceInstance
@@ -763,6 +777,7 @@ class ServiceInstance(InstanceResource):
             webhook_filters=webhook_filters,
             limits_channel_members=limits_channel_members,
             limits_user_channels=limits_user_channels,
+            media_compatibility_message=media_compatibility_message,
         )
 
     @property